backend/requirements.txt
# Core Dependencies
fastapi==0.116.1
sqlmodel==0.0.24
uvicorn==0.35.0
python-multipart==0.0.20
pyjwt==2.10.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
boto3==1.39.4
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic-settings==2.10.1
phonenumbers==9.0.9
pymupdf==1.26.3


# AI/ML Dependencies
instructor==1.9.2
openai==1.95.1
numpy==2.2.6
pillow==11.3.0
pymupdf==1.26.3
chromadb==1.0.15
sentence-transformers==5.0.0
